Rename `SongPage` -> `UserPage`
